// packages/form-types/src/form-step.type.d.ts
export declare const formStepSchema: z.ZodObject<
  {
    stepId: z.ZodString;
// ...
  },
  z.core.$strip
>;
// ...
export declare const recipeBlockFieldSchema: z.ZodObject<
  {
    ref: z.ZodString;
    overrides: z.ZodOptional<
      z.ZodRecord<
        z.ZodString,
        z.ZodObject<
          {
            options: z.ZodOptional<
              z.ZodArray<
                z.ZodObject<
                  {
                    label: z.ZodString;
                    value: z.ZodString;
                    disabled: z.ZodOptional<z.ZodBoolean>;
                  },
                  z.core.$strip
                >
              >
            >;
            label: z.ZodString;
            placeholder: z.ZodOptional<z.ZodString>;
            hint: z.ZodOptional<z.ZodString>;
            defaultValue: z.ZodOptional<z.ZodAny>;
            isDisabled: z.ZodOptional<z.ZodBoolean>;
            isHidden: z.ZodOptional<z.ZodBoolean>;
            validations: z.ZodOptional<
              z.ZodRecord<
                z.ZodEnum<{
                  required: "required";
                  pattern: "pattern";
                  maxLength: "maxLength";
                  minLength: "minLength";
                  maxItems: "maxItems";
                  minItems: "minItems";
                  equal: "equal";
                  notEqual: "notEqual";
                  min: "min";
                  max: "max";
                  conditionalOn: "conditionalOn";
                  past: "past";
                  pastOrToday: "pastOrToday";
                  future: "future";
                  futureOrToday: "futureOrToday";
                  after: "after";
                  before: "before";
                  onOrAfter: "onOrAfter";
                  onOrBefore: "onOrBefore";
                  minYear: "minYear";
                  maxYear: "maxYear";
                  radio: "radio";
                  minSelection: "minSelection";
                  maxSelection: "maxSelection";
                  email: "email";
                  fileTypes: "fileTypes";
                  itemMaxSize: "itemMaxSize";
                  maxSize: "maxSize";
                  gt: "gt";
                  lt: "lt";
                  contains: "contains";
                  strictEquality: "strictEquality";
                }> &
                  z.core.$partial,
                z.ZodObject<
                  {
                    error: z.ZodOptional<z.ZodString>;
                    value: z.ZodOptional<z.ZodAny>;
                    reference: z.ZodOptional<z.ZodString>;
                    targetStepId: z.ZodOptional<z.ZodString>;
                  },
                  z.core.$strip
                >
              >
            >;
            multiple: z.ZodOptional<z.ZodBoolean>;
          },
          z.core.$strip
        >
      >
    >;
  },
  z.core.$strip
>;